pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "playlist-generator"
version = "3.2.0"
description = "Playlist Generator with PySide6 GUI"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PySide6",
    "PyYAML",
    "platformdirs",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
